requirements.txt
scapy==2.4.4
paramiko==2.7.2
dnspython==2.1.0
dnslib==0.9.14
pyftpdlib==1.5.6
impacket==0.9.22